backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
pydantic==1.10.9
openai==1.33.0
requests==2.31.0
python-dotenv==1.0.0